requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.2
django-rest-framework==0.1.0
django-rest-framework-nested==0.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
fonttools==4.53.1
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
